# app/api/eval_tasks.py
from typing import List, Dict, Any
from fastapi import APIRouter, Depends, status
from app.api.auth import get_current_user
from app.schemas.task import EvalTaskBase, EvalTaskCreate, EvalTaskResponse, EvalTaskRunResponse, EvalSampleResultsResponse, EvalMetricItem
from app.services.task_service import TaskService

router = APIRouter()

@router.post("/", response_model=EvalTaskResponse)
def create_task(
    task_in: EvalTaskBase,
    current_user: Dict[str, Any] = Depends(get_current_user)
):
    """
    创建评测任务
    需指定模型和数据集
    """
    full_task_in = EvalTaskCreate(user_id=current_user["id"], **task_in.model_dump())
    return TaskService.create_task(full_task_in, current_user["id"])

@router.get("/", response_model=List[EvalTaskResponse])
def list_tasks(
    current_user: Dict[str, Any] = Depends(get_current_user)
):
    """
    获取用户下的评测任务列表
    """
    return TaskService.list_tasks(current_user["id"])

@router.get("/{task_id}", response_model=EvalTaskResponse)
def get_task(
    task_id: str,
    current_user: Dict[str, Any] = Depends(get_current_user)
):
    """
    获取任务详情
    """
    return TaskService.get_task(task_id, current_user["id"])

@router.post("/{task_id}/run", response_model=EvalTaskRunResponse)
async def run_task(
    task_id: str,
    current_user: Dict[str, Any] = Depends(get_current_user)
):
    """
    执行评测任务
    触发后台异步执行
    """
    return await TaskService.run_task(task_id, current_user["id"])

@router.get("/{task_id}/samples", response_model=EvalSampleResultsResponse)
def get_task_samples(
    task_id: str,
    page: int = 1,
    size: int = 50,
    current_user: Dict[str, Any] = Depends(get_current_user)
):
    return TaskService.get_task_samples(task_id, current_user["id"], page, size)

@router.get("/{task_id}/metrics", response_model=List[EvalMetricItem])
def get_task_metrics(
    task_id: str,
    current_user: Dict[str, Any] = Depends(get_current_user)
):
    return TaskService.get_task_metrics(task_id, current_user["id"])