# openapi.json
{"openapi": "3.1.0", "info": {"title": "Safety Platform", "version": "0.1.0"}, "paths": {"/api/auth/register": {"post": {"tags": ["auth"], "summary": "Register", "description": "用户注册接口\n接收邮箱、密码和姓名，创建新用户", "operationId": "register_api_auth_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response Register Api Auth Register Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "description": "用户登录接口 (OAuth2 兼容)\n接收 username (邮箱) 和 password，验证成功后返回 Access Token", "operationId": "login_api_auth_login_post", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_login_api_auth_login_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/users/me": {"get": {"tags": ["auth"], "summary": "Read Users Me", "description": "获取当前登录用户信息\n需要携带有效的 Bearer Token", "operationId": "read_users_me_api_auth_users_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response Read Users Me Api Auth Users Me Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/presets": {"get": {"tags": ["model_configs"], "summary": "List Model Presets", "description": "获取平台预置模型列表\n返回可用的模型预设（ID, 名称, 描述等）", "operationId": "list_model_presets_api_presets_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"additionalProperties": true, "type": "object"}, "type": "array", "title": "Response List Model Presets Api Presets Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/models/from-preset": {"post": {"tags": ["model_configs"], "summary": "Create Model From Preset", "description": "使用预置模型创建配置\n无需用户输入 API Key，系统自动注入", "operationId": "create_model_from_preset_api_models_from_preset_post", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "preset_id", "in": "query", "required": true, "schema": {"type": "string", "title": "Preset Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ModelConfigResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/models": {"get": {"tags": ["model_configs"], "summary": "List User Models", "description": "获取用户下的所有模型配置", "operationId": "list_user_models_api_models_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ModelConfigResponse"}, "type": "array", "title": "Response List User Models Api Models Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["model_configs"], "summary": "Create Model Config", "description": "创建模型配置 (自定义)\n将模型关联到指定项目，并设置相关参数", "operationId": "create_model_config_api_models_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ModelConfigBase"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ModelConfigResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/models/{model_id}": {"get": {"tags": ["model_configs"], "summary": "Get Model Config", "description": "获取单个模型配置详情", "operationId": "get_model_config_api_models__model_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "model_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Model Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ModelConfigResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["model_configs"], "summary": "Delete Model Config", "description": "删除模型配置\n需要项目访问权限", "operationId": "delete_model_config_api_models__model_id__delete", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "model_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Model Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/datasets/presets": {"get": {"tags": ["datasets"], "summary": "List Dataset Presets", "description": "获取平台预置数据集列表", "operationId": "list_dataset_presets_api_datasets_presets_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"additionalProperties": true, "type": "object"}, "type": "array", "title": "Response List Dataset Presets Api Datasets Presets Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/datasets/from-preset": {"post": {"tags": ["datasets"], "summary": "Create Dataset From Preset", "description": "从预置数据集创建", "operationId": "create_dataset_from_preset_api_datasets_from_preset_post", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "preset_id", "in": "query", "required": true, "schema": {"type": "string", "title": "Preset Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DatasetResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/datasets/": {"get": {"tags": ["datasets"], "summary": "List Datasets", "description": "获取用户下的所有数据集", "operationId": "list_datasets_api_datasets__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/DatasetResponse"}, "type": "array", "title": "Response List Datasets Api Datasets  Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["datasets"], "summary": "Create Dataset", "description": "注册新数据集", "operationId": "create_dataset_api_datasets__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/DatasetBase"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DatasetResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/datasets/upload": {"post": {"tags": ["datasets"], "summary": "Upload Dataset", "operationId": "upload_dataset_api_datasets_upload_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload_dataset_api_datasets_upload_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DatasetResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/datasets/{dataset_id}": {"get": {"tags": ["datasets"], "summary": "Get Dataset", "description": "获取单个数据集详情", "operationId": "get_dataset_api_datasets__dataset_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "dataset_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Dataset Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DatasetResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/datasets/{dataset_id}/samples": {"get": {"tags": ["datasets"], "summary": "Get Dataset Samples", "operationId": "get_dataset_samples_api_datasets__dataset_id__samples_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "dataset_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Dataset Id"}}, {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Page"}}, {"name": "size", "in": "query", "required": false, "schema": {"type": "integer", "default": 50, "title": "Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DatasetSamplesResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/eval-tasks/": {"get": {"tags": ["eval_tasks"], "summary": "List Tasks", "description": "获取用户下的评测任务列表", "operationId": "list_tasks_api_eval_tasks__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/EvalTaskResponse"}, "type": "array", "title": "Response List Tasks Api Eval Tasks  Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["eval_tasks"], "summary": "Create Task", "description": "创建评测任务\n需指定模型和数据集", "operationId": "create_task_api_eval_tasks__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvalTaskBase"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvalTaskResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/eval-tasks/{task_id}": {"get": {"tags": ["eval_tasks"], "summary": "Get Task", "description": "获取任务详情", "operationId": "get_task_api_eval_tasks__task_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Task Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvalTaskResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/eval-tasks/{task_id}/run": {"post": {"tags": ["eval_tasks"], "summary": "Run Task", "description": "执行评测任务\n触发后台异步执行", "operationId": "run_task_api_eval_tasks__task_id__run_post", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Task Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvalTaskRunResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/eval-tasks/{task_id}/samples": {"get": {"tags": ["eval_tasks"], "summary": "Get Task Samples", "operationId": "get_task_samples_api_eval_tasks__task_id__samples_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Task Id"}}, {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Page"}}, {"name": "size", "in": "query", "required": false, "schema": {"type": "integer", "default": 50, "title": "Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvalSampleResultsResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/eval-tasks/{task_id}/metrics": {"get": {"tags": ["eval_tasks"], "summary": "Get Task Metrics", "operationId": "get_task_metrics_api_eval_tasks__task_id__metrics_get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Task Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/EvalMetricItem"}, "title": "Response Get Task Metrics Api Eval Tasks  Task Id  Metrics Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/health": {"get": {"summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"Body_login_api_auth_login_post": {"properties": {"grant_type": {"anyOf": [{"type": "string", "pattern": "password"}, {"type": "null"}], "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Id"}, "client_secret": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "Body_login_api_auth_login_post"}, "Body_upload_dataset_api_datasets_upload_post": {"properties": {"file": {"type": "string", "format": "binary", "title": "File"}, "name": {"type": "string", "title": "Name"}, "description": {"type": "string", "title": "Description"}, "schema_json": {"type": "string", "title": "Schema Json"}, "prompt_field": {"type": "string", "title": "Prompt Field"}, "system_prompt": {"type": "string", "title": "System Prompt"}}, "type": "object", "required": ["file", "name"], "title": "Body_upload_dataset_api_datasets_upload_post"}, "DatasetBase": {"properties": {"name": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Name"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "source_type": {"type": "string", "title": "Source Type", "description": "file_upload, s3, url"}, "storage_uri": {"type": "string", "maxLength": 255, "title": "Storage Uri"}, "schema_json": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Schema Json"}}, "type": "object", "required": ["name", "source_type", "storage_uri"], "title": "DatasetBase"}, "DatasetResponse": {"properties": {"name": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Name"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "source_type": {"type": "string", "title": "Source Type", "description": "file_upload, s3, url"}, "storage_uri": {"type": "string", "maxLength": 255, "title": "Storage Uri"}, "schema_json": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Schema Json"}, "id": {"type": "string", "title": "Id"}, "user_id": {"type": "string", "title": "User Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["name", "source_type", "storage_uri", "id", "user_id", "created_at"], "title": "DatasetResponse"}, "DatasetSampleItem": {"properties": {"id": {"anyOf": [{}, {"type": "null"}], "title": "Id"}, "prompt": {"type": "string", "title": "Prompt"}, "data": {"title": "Data"}}, "type": "object", "required": ["prompt", "data"], "title": "DatasetSampleItem"}, "DatasetSamplesResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/DatasetSampleItem"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}}, "type": "object", "required": ["items", "total"], "title": "DatasetSamplesResponse"}, "EvalMetricItem": {"properties": {"metric_name": {"type": "string", "title": "Metric Name"}, "metric_value": {"type": "number", "title": "Metric Value"}, "details_json": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Details Json"}}, "type": "object", "required": ["metric_name", "metric_value"], "title": "EvalMetricItem"}, "EvalSampleResultItem": {"properties": {"sample_id": {"type": "string", "title": "Sample Id"}, "input_text": {"type": "string", "title": "Input Text"}, "model_output": {"type": "string", "title": "Model Output"}, "labels_json": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Labels Json"}, "score_json": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Score Json"}}, "type": "object", "required": ["sample_id", "input_text", "model_output"], "title": "EvalSampleResultItem"}, "EvalSampleResultsResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/EvalSampleResultItem"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}}, "type": "object", "required": ["items", "total"], "title": "EvalSampleResultsResponse"}, "EvalTaskBase": {"properties": {"name": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Name"}, "model_config_id": {"type": "string", "title": "Model Config Id"}, "dataset_id": {"type": "string", "title": "Dataset Id"}, "attack_strategy_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Attack Strategy Id"}, "metric_set_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Metric Set Id"}, "task_type": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Task Type", "default": "hallucination"}}, "type": "object", "required": ["name", "model_config_id", "dataset_id"], "title": "EvalTaskBase"}, "EvalTaskResponse": {"properties": {"name": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Name"}, "model_config_id": {"type": "string", "title": "Model Config Id"}, "dataset_id": {"type": "string", "title": "Dataset Id"}, "attack_strategy_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Attack Strategy Id"}, "metric_set_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Metric Set Id"}, "task_type": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Task Type", "default": "hallucination"}, "id": {"type": "string", "title": "Id"}, "user_id": {"type": "string", "title": "User Id"}, "status": {"type": "string", "title": "Status"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "started_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Started At"}, "finished_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Finished At"}}, "type": "object", "required": ["name", "model_config_id", "dataset_id", "id", "user_id", "status", "created_at"], "title": "EvalTaskResponse"}, "EvalTaskRunResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "task_id": {"type": "string", "title": "Task Id"}, "run_no": {"type": "integer", "title": "Run No"}, "status": {"type": "string", "title": "Status"}, "started_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Started At"}, "finished_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Finished At"}, "error_message": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Error Message"}}, "type": "object", "required": ["id", "task_id", "run_no", "status"], "title": "EvalTaskRunResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ModelConfigBase": {"properties": {"name": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Name"}, "provider": {"type": "string", "maxLength": 50, "minLength": 1, "title": "Provider"}, "endpoint": {"type": "string", "maxLength": 255, "title": "Endpoint"}, "auth_type": {"type": "string", "maxLength": 20, "title": "Auth Type"}, "auth_secret_ref": {"anyOf": [{"type": "string", "maxLength": 255}, {"type": "null"}], "title": "Auth Secret Ref"}, "params_json": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Params Json"}}, "type": "object", "required": ["name", "provider", "endpoint", "auth_type"], "title": "ModelConfigBase"}, "ModelConfigResponse": {"properties": {"name": {"type": "string", "maxLength": 100, "minLength": 1, "title": "Name"}, "provider": {"type": "string", "maxLength": 50, "minLength": 1, "title": "Provider"}, "endpoint": {"type": "string", "maxLength": 255, "title": "Endpoint"}, "auth_type": {"type": "string", "maxLength": 20, "title": "Auth Type"}, "auth_secret_ref": {"anyOf": [{"type": "string", "maxLength": 255}, {"type": "null"}], "title": "Auth Secret Ref"}, "params_json": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Params Json"}, "id": {"type": "string", "title": "Id"}, "user_id": {"type": "string", "title": "User Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["name", "provider", "endpoint", "auth_type", "id", "user_id", "created_at"], "title": "ModelConfigResponse"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type"}}, "type": "object", "required": ["access_token", "token_type"], "title": "Token"}, "UserCreate": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "title": "Password"}, "name": {"type": "string", "title": "Name"}}, "type": "object", "required": ["email", "password", "name"], "title": "UserCreate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "/api/auth/login"}}}}}}